requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
PyNaCl>=1.5.0
click>=8.1.7
pytest>=8.0.0
pytest-mock>=3.12.0
pyinstaller>=6.0.0